+++ CHANGELOG.md
@@ -6,8 +6,8 @@ This project tries to follow [Semantic Versioning](http://semver.org/) since the
 
 This document mainly describes API changes important to users of this library.
 
-
 ## 1.2.x - Unreleased
+
 * Add static method `UriHelper::getDynamicStackFromStackConfig` and `UriHelper::getDynamicStackFromStackObject`
 
 ## 1.1.0 - 2017-11-13